requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
pkg-resources==0.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
